requirements.txt
alabaster==1.0.0
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
docutils==0.22.4
executing==2.2.1
icecream==2.1.8
idna==3.11
imagesize==1.4.1
ipdb==0.13.13
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.2.1
pip-licenses==5.5.0
pipdeptree==2.30.0
prettytable==3.17.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestats==0.5.0
pygame==2.6.1
Pygments==2.19.2
referencing==0.37.0
requests==2.32.5
roman-numerals==4.1.0
rpds-py==0.30.0
snowballstemmer==3.0.1
Sphinx==9.0.4
sphinx-autodoc-typehints==3.6.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.6.2
wcwidth==0.2.14
